# api-definitions/ai-openapi.json
{
  "paths": {
...
    "/addresses/multiuprn": {
      "post": {
        "operationId": "multiUprn",
        "tags": [
          "routes"
        ],
        "summary": "Gets address by list of UPRNs, posted as a JSON array called uprns.",
        "description": "Returns a single address for each UPRN in the array.",
        "parameters": [
          {
            "in": "query",
            "name": "historical",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": true
            },
            "description": "Include historical addresses"
          },
          {
            "in": "query",
            "name": "verbose",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false
            },
            "description": "Include the full address details in the response (including relatives, crossRefs, paf and nag)."
          },
          {
            "in": "query",
            "name": "epoch",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "description": "Select a specific AddressBase Epoch to search (latest if not supplied)."
          },
          {
            "in": "query",
            "name": "pafdefault",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": false
            },
            "description": "Populate the formatted address field with PAF if present"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. A json return of matched addresses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.uprn.AddressByMultiUprnResponseContainer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Indicates an issue with the request. Further details are provided in the response."
          },
          "401": {
            "description": "Unauthorised. The API key provided with the request is invalid."
          },
          "404": {
            "description": "Multi-uprn skips not found UPRNs, so does not normall return 404s"
          },
          "429": {
            "description": "Server too busy. The Address Index API is experiencing exceptional load."
          },
          "500": {
            "description": "Internal server error. Failed to process the request due to an internal error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "uk.gov.ons.addressIndex.model.server.response.uprn.AddressByMultiUprnResponse": {
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS"
            }
          },
          "historical": {
            "type": "boolean"
          },
          "epoch": {
            "type": "string"
          },
          "verbose": {
            "type": "boolean"
          },
          "pafdefault": {
            "type": "boolean"
          }
        },
        "required": [
          "addresses",
          "historical",
          "epoch",
          "verbose",
          "pafdefault"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.uprn.AddressByMultiUprnResponseContainer": {
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "dataVersion": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.uprn.AddressByMultiUprnResponse"
          },
          "status": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError"
            }
          }
        },
        "required": [
          "apiVersion",
          "dataVersion",
          "response",
          "status",
          "errors"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseGeo": {
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "easting": {
            "type": "integer",
            "format": "int32"
          },
          "northing": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "easting",
          "northing"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseSao": {
        "properties": {
          "saoText": {
            "type": "string"
          },
          "saoStartNumber": {
            "type": "string"
          },
          "saoStartSuffix": {
            "type": "string"
          },
          "saoEndNumber": {
            "type": "string"
          },
          "saoEndSuffix": {
            "type": "string"
          }
        },
        "required": [
          "saoText",
          "saoStartNumber",
          "saoStartSuffix",
          "saoEndNumber",
          "saoEndSuffix"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePao": {
        "properties": {
          "paoText": {
            "type": "string"
          },
          "paoStartNumber": {
            "type": "string"
          },
          "paoStartSuffix": {
            "type": "string"
          },
          "paoEndNumber": {
            "type": "string"
          },
          "paoEndSuffix": {
            "type": "string"
          }
        },
        "required": [
          "paoText",
          "paoStartNumber",
          "paoStartSuffix",
          "paoEndNumber",
          "paoEndSuffix"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseNag": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "postcodeLocator": {
            "type": "string"
          },
          "addressBasePostal": {
            "type": "string"
          },
          "usrn": {
            "type": "string"
          },
          "lpiKey": {
            "type": "string"
          },
          "pao": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePao"
          },
          "sao": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseSao"
          },
          "level": {
            "type": "string"
          },
          "officialFlag": {
            "type": "string"
          },
          "logicalStatus": {
            "type": "string"
          },
          "streetDescriptor": {
            "type": "string"
          },
          "townName": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "organisation": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "localCustodianCode": {
            "type": "string"
          },
          "localCustodianName": {
            "type": "string"
          },
          "localCustodianGeogCode": {
            "type": "string"
          },
          "lpiEndDate": {
            "type": "string"
          },
          "lpiStartDate": {
            "type": "string"
          }
        },
        "required": [
          "uprn",
          "postcodeLocator",
          "addressBasePostal",
          "usrn",
          "lpiKey",
          "pao",
          "sao",
          "level",
          "officialFlag",
          "logicalStatus",
          "streetDescriptor",
          "townName",
          "locality",
          "organisation",
          "legalName",
          "localCustodianCode",
          "localCustodianName",
          "localCustodianGeogCode",
          "lpiEndDate",
          "lpiStartDate"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePaf": {
        "properties": {
          "udprn": {
            "type": "string"
          },
          "organisationName": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "subBuildingName": {
            "type": "string"
          },
          "buildingName": {
            "type": "string"
          },
          "buildingNumber": {
            "type": "string"
          },
          "dependentThoroughfare": {
            "type": "string"
          },
          "thoroughfare": {
            "type": "string"
          },
          "doubleDependentLocality": {
            "type": "string"
          },
          "dependentLocality": {
            "type": "string"
          },
          "postTown": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "postcodeType": {
            "type": "string"
          },
          "deliveryPointSuffix": {
            "type": "string"
          },
          "welshDependentThoroughfare": {
            "type": "string"
          },
          "welshThoroughfare": {
            "type": "string"
          },
          "welshDoubleDependentLocality": {
            "type": "string"
          },
          "welshDependentLocality": {
            "type": "string"
          },
          "welshPostTown": {
            "type": "string"
          },
          "poBoxNumber": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "udprn",
          "organisationName",
          "departmentName",
          "subBuildingName",
          "buildingName",
          "buildingNumber",
          "dependentThoroughfare",
          "thoroughfare",
          "doubleDependentLocality",
          "dependentLocality",
          "postTown",
          "postcode",
          "postcodeType",
          "deliveryPointSuffix",
          "welshDependentThoroughfare",
          "welshThoroughfare",
          "welshDoubleDependentLocality",
          "welshDependentLocality",
          "welshPostTown",
          "poBoxNumber",
          "startDate",
          "endDate"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlightHit": {
        "properties": {
          "source": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "distinctHitCount": {
            "type": "integer",
            "format": "int32"
          },
          "highLightedText": {
            "type": "string"
          }
        },
        "required": [
          "source",
          "lang",
          "distinctHitCount",
          "highLightedText"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight": {
        "properties": {
          "bestMatchAddress": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlightHit"
            }
          }
        },
        "required": [
          "bestMatchAddress",
          "source",
          "lang"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef": {
        "properties": {
          "crossReference": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "crossReference",
          "source"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative": {
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "siblings": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "parents": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "required": [
          "level",
          "siblings",
          "parents"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "parentUprn": {
            "type": "string"
          },
          "relatives": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative"
            }
          },
          "crossRefs": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef"
            }
          },
          "formattedAddress": {
            "type": "string"
          },
          "formattedAddressNag": {
            "type": "string"
          },
          "formattedAddressPaf": {
            "type": "string"
          },
          "welshFormattedAddressNag": {
            "type": "string"
          },
          "welshFormattedAddressPaf": {
            "type": "string"
          },
          "highlights": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight"
          },
          "paf": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePaf"
          },
          "nag": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseNag"
            }
          },
          "geo": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseGeo"
          },
          "classificationCode": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "lpiLogicalStatus": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number",
            "format": "double"
          },
          "underlyingScore": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "uprn",
          "parentUprn",
          "formattedAddress",
          "formattedAddressNag",
          "formattedAddressPaf",
          "welshFormattedAddressNag",
          "welshFormattedAddressPaf",
          "classificationCode",
          "countryCode",
          "lpiLogicalStatus",
          "confidenceScore",
          "underlyingScore"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
    }
  },
...
  "openapi": "3.0.0",
  "info": {
    "title": "ONS Address Index API",
    "description": "The ONS Address Index API lets you retrieve the official version of addresses in England and Wales. The data is currently updated on a 6 week rolling basis.",
    "version": "see /version below"
  }
}